Revert "[Config Refactor] Derive deploy override fields from stage config (#3…"
This reverts commit 01ebc0cdf542bb46db90ca8ecb641f26ca9247a0.
