+++ tests/test_config_factory.py
@@ -93,7 +93,6 @@ class TestStageConfig:
         assert omega_config.engine_args.worker_type == "ar"
         assert omega_config.final_output is True
         assert omega_config.final_output_type == "text"
-        assert "max_num_seqs" not in omega_config.engine_args
         # Legacy field name for backward compatibility
         assert omega_config.engine_input_source == []
 
@@ -146,24 +145,6 @@ class TestStageConfig:
         )
         omega_config = config.to_omegaconf()
         assert omega_config.engine_args.max_num_seqs == 32
-
-    def test_to_omegaconf_omits_none_deploy_overrides_for_engine_args(self):
-        """None deploy overrides must fall through to EngineArgs defaults."""
-        from vllm_omni.config.stage_config import deploy_override_field_names
-
-        config = StageConfig(
-            stage_id=0,
-            model_stage="thinker",
-            runtime_overrides={name: None for name in deploy_override_field_names()},
-        )
-
-        omega_config = config.to_omegaconf()
-        engine_args = dict(omega_config.engine_args)
-
-        assert "devices" not in engine_args
-        assert "max_batch_size" not in engine_args
-        for name in deploy_override_field_names() - {"devices"}:
-            assert name not in engine_args
 
 
 class TestModelPipeline:
@@ -821,40 +802,6 @@ class TestPipelineRegistry:
 
 
 class TestDeployConfigLoading:
-    def test_deploy_override_fields_include_deploy_schema_fields(self):
-        from vllm_omni.config.stage_config import deploy_override_field_names
-
-        expected_fields = {
-            "async_chunk",
-            "async_scheduling",
-            "config_format",
-            "data_parallel_size",
-            "devices",
-            "disable_hybrid_kv_cache_manager",
-            "distributed_executor_backend",
-            "dtype",
-            "enable_chunked_prefill",
-            "enable_flashinfer_autotune",
-            "enable_prefix_caching",
-            "enforce_eager",
-            "gpu_memory_utilization",
-            "load_format",
-            "max_model_len",
-            "max_num_batched_tokens",
-            "max_num_seqs",
-            "mm_processor_cache_gb",
-            "pipeline_parallel_size",
-            "profiler_config",
-            "quantization",
-            "skip_mm_profiling",
-            "subtalker_sampling_params",
-            "tensor_parallel_size",
-            "tokenizer_mode",
-            "trust_remote_code",
-        }
-
-        assert expected_fields == deploy_override_field_names()
-
     def test_load_deploy_config(self):
         from pathlib import Path
 
@@ -869,17 +816,6 @@ class TestDeployConfigLoading:
         assert deploy.async_chunk is True
         assert deploy.connectors is not None
         assert deploy.platforms is not None
-
-        voxtral_path = Path(__file__).parent.parent / "vllm_omni" / "deploy" / "voxtral_tts.yaml"
-        if voxtral_path.exists():
-            voxtral_deploy = load_deploy_config(voxtral_path)
-            assert voxtral_deploy.stages[0].config_format == "mistral"
-            assert voxtral_deploy.stages[0].load_format == "mistral"
-            assert voxtral_deploy.stages[0].tokenizer_mode == "mistral"
-            assert not any(
-                name in voxtral_deploy.stages[0].engine_extras
-                for name in ("config_format", "load_format", "tokenizer_mode")
-            )
 
     def test_merge_pipeline_deploy(self):
         from pathlib import Path
@@ -1075,8 +1011,7 @@ class TestBaseConfigInheritance:
         deploy = load_deploy_config(ci_path)
         assert len(deploy.stages) == 3
         # CI overrides
-        assert deploy.stages[0].load_format == "dummy"
-        assert "load_format" not in deploy.stages[0].engine_extras
+        assert deploy.stages[0].engine_extras.get("load_format") == "dummy"
         assert deploy.stages[0].max_num_seqs == 5
         # Inherited from base
         assert deploy.stages[0].gpu_memory_utilization == 0.9
@@ -1281,7 +1216,7 @@ class TestSentinelDefaultPrecedence:
     def test_none_value_skipped_yaml_wins(self):
         stages = self._stages({"max_num_seqs": None})
         assert stages[2].runtime_overrides.get("max_num_seqs") is None
-        assert "max_num_seqs" not in stages[2].yaml_engine_args
+        assert stages[2].yaml_engine_args.get("max_num_seqs") == 1
 
     def test_empty_kwargs_yaml_only(self):
         stages = self._stages({})
