+++ vllm_omni/deploy/cosyvoice3.yaml
@@ -40,7 +40,7 @@ stages:
       # near-identity repetition penalty forces vLLM to track
       # output_token_ids for RAS (stop-token logit logsumexp).
       repetition_penalty: 1.0001
-    enable_prefix_caching: false
+    disable_hybrid_kv_cache_manager: true
     mm_processor_cache_gb: 0
     skip_mm_profiling: true
 
@@ -54,5 +54,5 @@ stages:
       from_stage_0: connector_of_shared_memory
     default_sampling_params:
       max_tokens: 2048
-    enable_prefix_caching: false
+    disable_hybrid_kv_cache_manager: true
     skip_mm_profiling: true
