+++ vllm_omni/deploy/qwen3_omni_moe.yaml
@@ -23,7 +23,6 @@ connectors:
 stages:
   - stage_id: 0
     gpu_memory_utilization: 0.9
-    enable_prefix_caching: false
     devices: "0"
     default_sampling_params:
       temperature: 0.4
@@ -48,7 +47,6 @@ stages:
   - stage_id: 2
     gpu_memory_utilization: 0.1
     enforce_eager: true
-    enable_prefix_caching: false
     async_scheduling: false
     max_num_batched_tokens: 51200
     devices: "1"
